requirements.txt
beautifulsoup4==4.13.4
certifi==2024.6.2
charset-normalizer==3.4.2
idna==3.10
joblib==1.5.1
numpy==1.26.4
requests==2.32.3
scikit-learn==1.6.1
scipy==1.13.0
soupsieve==2.7
tabulate==0.9.0
threadpoolctl==3.6.0
typing_extensions==4.13.2
urllib3==2.4.0
psycopg2-binary
python-dotenv
fastapi
uvicorn
